+++ CHANGELOG.md
@@ -7,11 +7,13 @@ We follow [Semantic Versions](https://semver.org/).
 
 ### Features
 
+- Added `@attrs.define` official support, #706
 - Added `MediaType` validation for the default `encoding` field
   and OpenAPI 3.2 `itemEncoding` and `prefixEncoding` fields, #695
 - Added `MediaTypeMetadata` metadata item to set required parameters
   for the `MediaType` request body parameters
   for `Body` and `FileMedata` components, #695 and #698
+- Added support for Swagger, Redoc, and Scalar CDN configuration, #678
 
 ### Bugfixes